radar queries
